move ModuleDecoratorDependency into a separate file
