+++ lib/NodeStuffPlugin.js
@@ -6,7 +6,6 @@
 "use strict";
 
 const path = require("path");
-const InitFragment = require("./InitFragment");
 const {
 	evaluateToIdentifier,
 	evaluateToString,
@@ -16,7 +15,7 @@ const {
 	toConstantDependencyWithWebpackRequire
 } = require("./JavascriptParserHelpers");
 const CachedConstDependency = require("./dependencies/CachedConstDependency");
-const ModuleDependency = require("./dependencies/ModuleDependency");
+const ModuleDecoratorDependency = require("./dependencies/ModuleDecoratorDependency");
 
 /** @typedef {import("webpack-sources").ReplaceSource} ReplaceSource */
 /** @typedef {import("./Dependency")} Dependency */
@@ -42,7 +41,7 @@ class NodeStuffPlugin {
 				);
 				compilation.dependencyTemplates.set(
 					ModuleDecoratorDependency,
-					new ModuleDecoratorDependencyTemplate()
+					new ModuleDecoratorDependency.Template()
 				);
 
 				const handler = (parser, parserOptions) => {
@@ -189,49 +188,6 @@ class NodeStuffPlugin {
 					.tap("NodeStuffPlugin", handler);
 			}
 		);
-	}
-}
-
-class ModuleDecoratorDependency extends ModuleDependency {
-	constructor(request, originalModule, range) {
-		super(request);
-		this.originalModule = originalModule;
-		this.range = range;
-	}
-}
-
-class ModuleDecoratorDependencyTemplate extends ModuleDependency.Template {
-	/**
-	 * @param {Dependency} dependency the dependency for which the template should be applied
-	 * @param {ReplaceSource} source the current replace source which can be modified
-	 * @param {RuntimeTemplate} runtimeTemplate the runtime template
-	 * @param {DependencyTemplates} dependencyTemplates the dependency templates
-	 * @returns {void}
-	 */
-	apply(dependency, source, runtimeTemplate, dependencyTemplates) {}
-
-	/**
-	 * @param {Dependency} dependency the dependency for which the template should be applied
-	 * @param {ReplaceSource} source the current replace source which can be modified
-	 * @param {RuntimeTemplate} runtimeTemplate the runtime template
-	 * @param {DependencyTemplates} dependencyTemplates the dependency templates
-	 * @returns {InitFragment[]|null} the init fragments
-	 */
-	getInitFragments(dependency, source, runtimeTemplate, dependencyTemplates) {
-		const dep = /** @type {ModuleDecoratorDependency} */ (dependency);
-		return [
-			new InitFragment(
-				`/* module decorator */ ${
-					dep.originalModule.moduleArgument
-				} = ${runtimeTemplate.moduleExports({
-					module: dep.module,
-					request: dep.request
-				})}(${dep.originalModule.moduleArgument});\n`,
-				InitFragment.STAGE_PROVIDES,
-				0,
-				`module decorator ${dep.originalModule.id}`
-			)
-		];
 	}
 }
